refactor: `browser` api usage
- Removed `webextension-polyfill`
